Package

+++ recipes/default.rb
@@ -3,32 +3,40 @@
 Erubis::Context.send(:include, Extensions::Templates)
 
 elasticsearch = "elasticsearch-#{node.elasticsearch[:version]}"
+include_recipe "elasticsearch::_default"
 
-include_recipe "elasticsearch::curl"
-include_recipe "ark"
+if node.elasticsearch[:method] == "pkg"
+  case node['platform']
+    when "ubuntu","debian"
+      include_recipe "elasticsearch::deb"
+    when "redhat","fedora"
+      include_recipe "elasticsearch::rpm" 
+  end
+else
+  include_recipe "elasticsearch::curl"
+  include_recipe "ark"
 
-# Create user and group
-#
-group node.elasticsearch[:user] do
-  action :create
-end
+  group node.elasticsearch[:user] do
+    action :create
+  end
 
-user node.elasticsearch[:user] do
-  comment "ElasticSearch User"
-  home    "#{node.elasticsearch[:dir]}/elasticsearch"
-  shell   "/bin/bash"
-  gid     node.elasticsearch[:user]
-  supports :manage_home => false
-  action  :create
-end
+  user node.elasticsearch[:user] do
+    comment "ElasticSearch User"
+    home    "#{node.elasticsearch[:dir]}/elasticsearch"
+    shell   "/bin/bash"
+    gid     node.elasticsearch[:user]
+    supports :manage_home => false
+    action  :create
+  end
 
-# FIX: Work around the fact that Chef creates the directory even for `manage_home: false`
-bash "remove the elasticsearch user home" do
-  user    'root'
-  code    "rm -rf  #{node.elasticsearch[:dir]}/elasticsearch"
+  # FIX: Work around the fact that Chef creates the directory even for `manage_home: false`
+  bash "remove the elasticsearch user home" do
+    user    'root'
+    code    "rm -rf  #{node.elasticsearch[:dir]}/elasticsearch"
 
-  not_if  "test -L #{node.elasticsearch[:dir]}/elasticsearch"
-  only_if "test -d #{node.elasticsearch[:dir]}/elasticsearch"
+    not_if  "test -L #{node.elasticsearch[:dir]}/elasticsearch"
+    only_if "test -d #{node.elasticsearch[:dir]}/elasticsearch"
+  end
 end
 
 # Create ES directories
@@ -60,35 +68,39 @@ template "/etc/init.d/elasticsearch" do
   owner 'root' and mode 0755
 end
 
-service "elasticsearch" do
-  supports :status => true, :restart => true
-  action [ :enable ]
-end
 
-# Download, extract, symlink the elasticsearch libraries and binaries
-#
-ark_prefix_root = node.elasticsearch[:dir] || node.ark[:prefix_root]
-ark_prefix_home = node.elasticsearch[:dir] || node.ark[:prefix_home]
 
-ark "elasticsearch" do
-  url   node.elasticsearch[:download_url]
-  owner node.elasticsearch[:user]
-  group node.elasticsearch[:user]
-  version node.elasticsearch[:version]
-  has_binaries ['bin/elasticsearch', 'bin/plugin']
-  checksum node.elasticsearch[:checksum]
-  prefix_root   ark_prefix_root
-  prefix_home   ark_prefix_home
 
-  notifies :start,   'service[elasticsearch]'
-  notifies :restart, 'service[elasticsearch]'
+if node.elasticsearch[:method] == "source"
+  service "elasticsearch" do
+    supports :status => true, :restart => true
+    action [ :enable ]
+  end
+  # Download, extract, symlink the elasticsearch libraries and binaries
+  #
+  ark_prefix_root = node.elasticsearch[:dir] || node.ark[:prefix_root]
+  ark_prefix_home = node.elasticsearch[:dir] || node.ark[:prefix_home]
 
-  not_if do
-    link   = "#{node.elasticsearch[:dir]}/elasticsearch"
-    target = "#{node.elasticsearch[:dir]}/elasticsearch-#{node.elasticsearch[:version]}"
-    binary = "#{target}/bin/elasticsearch"
+  ark "elasticsearch" do
+    url   node.elasticsearch[:download_url]
+    owner node.elasticsearch[:user]
+    group node.elasticsearch[:user]
+    version node.elasticsearch[:version]
+    has_binaries ['bin/elasticsearch', 'bin/plugin']
+    checksum node.elasticsearch[:checksum]
+    prefix_root   ark_prefix_root
+    prefix_home   ark_prefix_home
 
-    ::File.directory?(link) && ::File.symlink?(link) && ::File.readlink(link) == target && ::File.exists?(binary)
+    notifies :start,   'service[elasticsearch]'
+    notifies :restart, 'service[elasticsearch]'
+
+    not_if do
+      link   = "#{node.elasticsearch[:dir]}/elasticsearch"
+      target = "#{node.elasticsearch[:dir]}/elasticsearch-#{node.elasticsearch[:version]}"
+      binary = "#{target}/bin/elasticsearch"
+
+      ::File.directory?(link) && ::File.symlink?(link) && ::File.readlink(link) == target && ::File.exists?(binary)
+    end
   end
 end
